+++ components/form/types/yes_no.type.tsx
@@ -8,7 +8,10 @@ export const YesNoType: React.FC<FieldTypeProps> = ({ field }) => {
       <Form.Item
         name={[field.id, 'value']}
         rules={[{ required: field.required, message: 'Please provide Information' }]}
-        initialValue={field.value}
+        initialValue={!!field.value}
+        valuePropName={'checked'}
+        getValueFromEvent={(checked: boolean) => (checked ? '1' : '')}
+        getValueProps={(e: string) => ({ checked: !!e })}
       >
         <Switch />
       </Form.Item>